+++ src/ampworks/gitt/_extract_params.py
@@ -103,7 +103,7 @@ def extract_params(data: Dataset, radius: float, tmin: float = 1,
 
     Examples
     --------
-    >>> data = amp.datasets.load_datasets('gitt_discharge')
+    >>> data = amp.datasets.load_datasets('gitt/gitt_discharge')
     >>> params, stats = amp.gitt.extract_params(data, 1.8e-6, return_all=True)
     >>> params.plot('SOC', 'Eeq')
     >>> params.plot('SOC', 'Ds', logy=True)